cli: avoid running Example_demo_locality in the race detector
See the discussion on #41045: running a 9-node demo cluster under the
race detector results in large CPU and ram usage that exceeds the CI
budget.

Since the various moving pieces in the code are race-tested in other
places, we are solely interested here to unit test functionality, and
it's not necessary to run in testrace.

Release justification: removes test flakiness

Release note: None
